refactor: remove WordPress backend integration
Blog content is now fully local MDX, so drop the WPGraphQL fetcher,
HTML post renderer, GraphQL types, WordPress environment variables,
CI secrets, remote image pattern and unused GraphQL dependencies.
