+++ .env.example
@@ -1,7 +1,3 @@
-WORDPRESS_API_URL=
-WORDPRESS_AUTH_REFRESH_TOKEN=
-WORDPRESS_PREVIEW_SECRET=
-
 AKISMET_API_KEY=
 
 MJ_APIKEY_PUBLIC=
