+++ .github/workflows/cd.yaml
@@ -50,7 +50,6 @@ jobs:
       - name: Build Docker images
         uses: docker/bake-action@v6
         env:
-          WORDPRESS_API_URL: ${{ secrets.WORDPRESS_API_URL }}
           NEXT_PUBLIC_SENTRY_RELEASE: ${{ env.NEXT_PUBLIC_SENTRY_RELEASE }}
         with:
           pull: true
