+++ .github/workflows/ci.yml
@@ -48,7 +48,6 @@ jobs:
       - name: Build Docker images
         uses: docker/bake-action@v6
         env:
-          WORDPRESS_API_URL: ${{ secrets.WORDPRESS_API_URL }}
           NEXT_PUBLIC_FRONT_URL: http://localhost:3000
           NEXT_PUBLIC_GTM_ID: GTM-XXXXX
           NEXT_PUBLIC_SENTRY_RELEASE: ci-build
