--- app/components/blog/PostBody.module.css
@@ -1,137 +1,0 @@
-@reference "../../(pages)/globals.css";
-
-.content {
-  @apply max-w-5xl w-full mx-auto;
-}
-
-.content > p:first-child {
-  @apply text-white text-2xl md:text-3xl;
-}
-
-.content [class^='wp-block-image'] {
-  @apply mb-4;
-}
-
-.content [class^='wp-block-image'] [class^='aligncenter'] {
-  @apply mx-auto table;
-}
-
-.content img {
-  @apply max-w-full h-auto;
-}
-
-.content {
-  @apply text-xl;
-}
-
-.content a {
-  @apply text-orange;
-}
-
-.content p > strong {
-  @apply text-white;
-}
-
-.content pre {
-  @apply -mx-4 rounded-lg bg-gray-darker p-4 border-2 border-[#4b2809] overflow-x-auto text-sm;
-}
-
-.content pre code {
-  @apply text-[#727272] text-sm;
-}
-
-.content code {
-  overflow-wrap: break-word;
-}
-
-.content h2,
-.content h3 {
-  @apply text-white font-bold leading-snug;
-}
-
-.content h2 {
-  @apply text-3xl mt-8 mb-6;
-}
-
-.content h3 {
-  @apply text-2xl mt-6 mb-4;
-}
-
-.content p,
-.content ul,
-.content ol,
-.content blockquote {
-  @apply my-4;
-}
-
-.content ol {
-  @apply list-decimal ml-8;
-}
-
-.content [class~='alignwide'] {
-  @apply -mx-4;
-}
-
-.content [class~='alignwide'] img {
-  @apply w-full;
-}
-
-.content p > code {
-  padding: 3px 5px;
-  font-size: 0.8em;
-  background: #141414;
-  color: #fff;
-  font-family:
-    Monaco,
-    Consolas,
-    Andale Mono,
-    DejaVu Sans Mono,
-    monospace;
-}
-
-.content [class~='wp-block-columns'] {
-  @apply block md:flex;
-}
-
-.content [class~='is-vertically-aligned-center'] {
-  @apply self-center;
-}
-
-.content [class~='wp-block-columns'] [class~='wp-block-column']:first-child {
-  @apply mr-8;
-}
-
-.content [class~='button'] {
-  @apply text-white;
-}
-
-.content [class~='has-text-align-center'] {
-  @apply text-center;
-}
-
-.content [class~='wp-block-gallery'] [class~='blocks-gallery-grid'] {
-  @apply grid gap-6;
-}
-
-.content
-  [class~='wp-block-gallery'][class~='columns-3']
-  [class~='blocks-gallery-grid'] {
-  @apply grid-cols-1 md:grid-cols-3;
-}
-
-.content [class~='wp-block-jetpack-gif-wrapper'] {
-  height: 0;
-  margin: 0;
-  padding: calc(56.2% + 12px) 0 0;
-  position: relative;
-  width: 100%;
-}
-
-.content [class~='wp-block-jetpack-gif-wrapper'] iframe {
-  border: 0;
-  height: 100%;
-  left: 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-}
