+++ app/types/environment.d.ts
@@ -2,8 +2,6 @@ declare global {
   namespace NodeJS {
     interface ProcessEnv {
       NODE_ENV: 'development' | 'production'
-      BACKEND_URL: string
-      WORDPRESS_API_URL: string
       AKISMET_API_KEY: string
 
       // Mailjet API
