--- app/types/graphql/portfolio.ts
@@ -1,6 +1,0 @@
-export type EntriesType = Array<{
-  id: number
-  name: string
-  excerpt: string
-  values: Array<string | boolean>
-}>
