--- app/types/graphql/sitemap-type.ts
@@ -1,16 +1,0 @@
-interface Item {
-  edges: Array<{
-    node: {
-      title: string
-      uri: string
-    }
-  }>
-}
-
-export interface Sitemap {
-  data: {
-    projets: Item
-    posts: Item
-    supports: Item
-  }
-}
