+++ package.json
@@ -19,10 +19,7 @@
     "@next/third-parties": "16.0.7",
     "@sentry/nextjs": "^10.29.0",
     "@types/mdx": "^2.0.13",
-    "@types/prismjs": "^1.26.5",
     "akismet-api": "^6.0.0",
-    "graphql": "^16.12.0",
-    "graphql-tag": "^2.12.6",
     "gsap": "^3.13.0",
     "import-in-the-middle": "^2.0.0",
     "moment": "^2.30.1",
@@ -30,9 +27,7 @@
     "next": "^16.0.7",
     "next-sitemap": "^4.2.3",
     "postcss": "^8.5.6",
-    "prismjs": "^1.30.0",
     "prop-types": "^15.8.1",
-    "raw-loader": "^4.0.2",
     "react": "19.2.1",
     "react-dom": "19.2.1",
     "react-hook-form": "^7.68.0",
@@ -42,7 +37,6 @@
     "request-ip": "^3.3.0",
     "sanitize-html": "^2.17.0",
     "sharp": "^0.34.5",
-    "swr": "^2.3.7",
     "ts-node": "^10.9.2"
   },
   "devDependencies": {
